// UnrealEngine52/Engine/Source/Programs/Horde/HordeDashboard/src/backend/Api.ts
export enum Priority {

	// Lowest priority
	Lowest = "Lowest",

	// Below normal priority
	BelowNormal = "BelowNormal",

	// Normal priority
	Normal = "Normal",

	// Above normal priority
	AboveNormal = "AboveNormal",

	// High priority
	High = "High",

	// Highest priority
	Highest = "Highest"
}
// ...
export enum JobStepBatchState {

	// Waiting for dependencies of at least one jobstep to complete
	Waiting = "Waiting",

	// Getting ready to execute
	Starting = "Starting",

	// Ready to execute
	Ready = "Ready",

	// Currently running
	Running = "Running",

	// Getting ready to execute
	Stopping = "Stopping",

	// All steps have finished executing
	Complete = "Complete"
}

/**Error code for a batch not being executed */
export enum JobStepBatchError {

	/** No error */
	None = "None",

	/** The stream for this job is unknown */
	UnknownStream = "UnknownStream",

	/** The given agent type for this batch was not valid for this stream */
	UnknownAgentType = "UnknownAgentType",

	/** The pool id referenced by the agent type was not found */
	UnknownPool = "UnknownPool",

	/** There are no agents in the given pool currently online */
	NoAgentsInPool = "NoAgentsInPool",

	/** There are no agents in this pool that are onlinbe */
	NoAgentsOnline = "NoAgentsOnline",

	/** Unknown workspace referenced by the agent type */
	UnknownWorkspace = "UnknownWorkspace",

	/** Cancelled */
	Cancelled = "Cancelled",

	/** Lost connection with the agent machine */
	LostConnection = "LostConnection",

	/** Lease terminated prematurely */
	Incomplete = "Incomplete",

	/** An error ocurred while executing the lease. Cannot be retried. */
	ExecutionError = "ExecutionError",

	/** The change that the job is running against is invalid. */
	UnknownShelf = "UnknownShelf"

}
// ...
export enum ReportPlacement {
	/** On a panel of its own */
	Panel = "Panel",
	/** In the summary panel */
	Summary = "Summary"
}

/**Information about a report associated with a job*/
export type GetReportResponse = {

	// Report placement
	placement: ReportPlacement;

	/** Name of the report */
	name: string;

	/** The artifact id */
	artifactId: string;
}


/**Information about a job */
export type GetJobResponse = {

	/**Unique Id for the job */
	id: string;

	/**Unique id of the stream containing this job */
	streamId: string;

	/**Name of the job */
	name: string;

	/** The changelist number to build */
	change?: number;

	/** The preflight changelist number */
	preflightChange?: number;

	/** Description of the preflight */
	preflightDescription?: string;

	/** The template type */
	templateId?: string;

	/** Hash of the actual template data */
	templateHash?: string;

	/** Hash of the graph for this job */
	graphHash?: string;

	/** The user that started this job */
	startedByUserInfo?: GetThinUserInfoResponse;

	/** The user that started this job */
	abortedByUserInfo?: GetThinUserInfoResponse;

	/** The roles to impersonate when executing this job */
	roles: string[];

	/** Priority of the job */
	priority: Priority;

	/** Whether the change will automatically be submitted or not */
	autoSubmit?: boolean;

	/** The submitted changelist number */
	autoSubmitChange?: number;

	/**  Message produced by trying to auto-submit the change */
	autoSubmitMessage?: string;

	/**Time that the job was created */
	createTime: Date | string;

	/** The global job state */
	state: JobState;

	/** Array of jobstep batches */
	batches?: GetBatchResponse[];

	/**  List of labels */
	labels?: GetLabelStateResponse[];

	/** The default label, containing the state of all steps that are otherwise not matched. */
	defaultLabel?: GetDefaultLabelStateResponse;

	/** List of reports */
	reports?: GetReportResponse[];

	/**  Parameters for the job */
	arguments: string[];

	/**The last update time for this job*/
	updateTime: Date | string;

	/** Whether to update issues based on the outcome of this job */
	updateIssues?: boolean;

	/**  Custom permissions for this object */
	acl?: GetAclResponse;

}
// ...
export type GetStepResponse = {

	/**The unique id of the step */
	id: string;

	/**Index of the node which this jobstep is to execute */
	nodeIdx: number;

	/**Current state of the job step. This is updated automatically when runs complete. */
	state: JobStepState;

	/**Current outcome of the jobstep */
	outcome: JobStepOutcome;

	/**Error describing additional context for why a step failed to complete*/
	error: JobStepError;

	/** If the step has been requested to abort	*/
	abortRequested?: boolean;

	/* The user that requested the abort of this step */
	abortedByUserInfo?: GetThinUserInfoResponse;

	/* The user that retried this step */
	retriedByUserInfo?: GetThinUserInfoResponse;

	/**The log id for this step */
	logId?: string;

	/** Time at which the batch was ready (UTC). */
	readyTime?: Date | string;

	/**Time at which the run started (UTC). */
	startTime?: Date | string;

	/**Time at which the run finished (UTC) */
	finishTime: Date | string;

	/** List of reports */
	reports?: GetReportResponse[];

	/**User-defined properties for this jobstep. */
	properties: { [key: string]: string };

}
// ...
export type GetBatchResponse = {

	/**Unique id for this batch */
	id: string;

	/**The unique log file id */
	logId?: string;

	/**Index of the group being executed */
	groupIdx: number;

	/**The state of this batch */
	state: JobStepBatchState;

	/**Error code for this batch */
	error: JobStepBatchError;

	/**Steps within this run */
	steps: GetStepResponse[];

	/**The agent assigned to execute this group */
	agentId?: string;

	/** The USD rate of an agent hour */
	agentRate?: number;

	/**The agent session holding this lease */
	sessionId?: string;

	/**The lease that's executing this group */
	leaseId?: string;

	/**The priority of this batch */
	weightedPriority: number;

	/**Time at which the group started (UTC). */
	startTime?: Date | string;

	/**Time at which the group finished (UTC) */
	finishTime?: Date | string;

}
// ...
export type GetAclEntryResponse = {
	/**Names of the user or group */
	roles: string[];

	/**Array of actions to allow */
	allow: string[];
}

/**Information about an ACL */
export type GetAclResponse = {
	/**Entries to replace the existing ACL */
	entries: GetAclEntryResponse[];

	/**Whether to inherit permissions from the parent entity */
	inheritPermissions: boolean;
}
// ...
export enum LabelState {

	/** label is not currently being built (no required nodes are present)*/
	Unspecified = "Unspecified",

	/** Steps are still running */
	Running = "Running",

	/** All steps are complete */
	Complete = "Complete"
}

/// Outcome of an aggregate
export enum LabelOutcome {

	/** Aggregate is not currently being built */
	Unspecified = "Unspecified",

	/** A step dependency failed */
	Failure = "Failure",

	/** A dependency finished with warnings */
	Warnings = "Warnings",

	/** Successful */
	Success = "Success"
}


/** State of the job */
export enum JobState {

	/** Waiting for resources*/
	Waiting = "Waiting",

	/** Currently running one or more steps*/
	Running = "Running",

	/** All steps have completed */
	Complete = "Complete"
}
// ...
export type GetLabelStateResponse = {
	/**State of the label */
	state?: LabelState;

	/**Outcome of the label */
	outcome?: LabelOutcome;
}

/**Information about the default label (ie. with inlined list of nodes) */
export type GetDefaultLabelStateResponse = GetLabelStateResponse &
{
	/**List of nodes covered by default label */
	nodes: string[];
}
// ...
export type GetThinUserInfoResponse = {

	/**  Id of the user */
	id: string;

	/** Name of the user */
	name: string;

	/**  The user's email address */
	email: string;

}